ServerApp/requirements.txt
asgiref==3.8.1
Django==5.0
sqlparse==0.5.2
tzdata==2024.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.0